adding 2 CSV files for testing in a testing directory.
first-run.csv is the same as the sample CSV file but it has a GUID column and a source column mocked up per our enhancement request with Kezia's team.
second-run.csv is the same as first-run.csv but every other row is a duplicate of first-run.csv. The "Link-Text" also notes DUP- as a prefix for every entry to allow for easy testing.