Rewrite of the github insert script
Some extra files.
test-assessment.csv is the result of clicking "none of the above" on every section of the assesment.
test-assessmentsmall.csv is every 4th line in the test-assessment.csv file.
delete-testing-github.sh is a bash script that will destroy your github repo and then create a new empty one with the same name when it is finished.